requirements.txt
# TPMS Heat Exchanger Optimizer - Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Fluid properties
CoolProp>=6.4.0  # Fallback for REFPROP

# Visualization and GUI
matplotlib>=3.5.0  # For GUI visualizations
streamlit>=1.28.0  # Streamlit UI
plotly>=5.14.0  # 3D visualization

# Mesh generation and export
scikit-image>=0.19.0  # Marching cubes for mesh generation
numpy-stl>=3.0.0  # STL export

# Optional: REFPROP support (requires REFPROP installation)
# REFPROP-dll>=10.0  # Uncomment if you have REFPROP installed

